requirements.txt
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
PuLP==2.8.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
Werkzeug==3.1.6
WTForms==3.1.2
